+++ src/helpers/showNoGraphvizMessaage.ts
@@ -1,7 +1,8 @@
 import * as vscode from "vscode";
+import { localize } from "../localization";
 
-export function showNoGraphvizMessaage() {
-  vscode.window.showInformationMessage(
-    "Application graphviz need to be installed, please check the README.md"
-  );
+export function showNoGraphvizMessage() {
+    vscode.window.showWarningMessage(
+        localize("Application graphviz need to be installed, please check the README.md")
+    );
 }